update / reorg unit/live tests
