+++ tests/Azure.EntityServices.Tests/Table/EntityTableClientWritingTests.cs
@@ -2,6 +2,7 @@
 using Azure.EntityServices.Queries;
 using Azure.EntityServices.Tables;
 using Azure.EntityServices.Tables.Extensions;
+using Azure.EntityServices.Tests.Table;
 using Common.Samples;
 using Common.Samples.Models;
 using Common.Samples.Tools.Fakes;
@@ -16,18 +17,13 @@ using System.Threading.Tasks;
 namespace Azure.EntityServices.Table.Tests
 {
     [TestClass]
-    public class EntityTableClientTests
+    public class EntityTableClientWritingTests
     {
-        private readonly Action<EntityTableClientOptions> _commonOptions;
+        private readonly Action<EntityTableClientOptions> _defaultOptions;
 
-        public EntityTableClientTests()
+        public EntityTableClientWritingTests()
         {
-            _commonOptions = (EntityTableClientOptions options) =>
-            {
-                options.CreateTableIfNotExists = true;
-                options.TableName = $"{nameof(EntityTableClientTests)}{Guid.NewGuid():N}";
-                options.HandleTagMutation = true;
-            };
+            _defaultOptions = EntityTableClientCommon.DefaultOptions(nameof(EntityTableClientWritingTests));
         }
 
         [TestMethod]
@@ -37,7 +33,7 @@ namespace Azure.EntityServices.Table.Tests
             var person = persons.First();
 
             var entityTable = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString)
-                .Configure(options => _commonOptions(options), c =>
+                .Configure(options => _defaultOptions(options), c =>
             {
                 c.
                  SetPartitionKey(p => p.TenantId)
@@ -52,13 +48,171 @@ namespace Azure.EntityServices.Table.Tests
         }
 
         [TestMethod]
-        public async Task Should_Refresh_Tag_When_Value_Updated()
+        public async Task Should_Merge_Entity()
         {
             var persons = Fakers.CreateFakePerson().Generate(1);
             var person = persons.First();
 
             var entityTable = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString)
-                .Configure(options => _commonOptions(options), c =>
+                .Configure(options => _defaultOptions(options), c =>
+                {
+                    c.
+                     SetPartitionKey(p => p.TenantId)
+                    .SetRowKeyProp(p => p.PersonId)
+                    .AddTag(p => p.LastName)
+                    .AddTag(p => p.Created);
+                });
+            await entityTable.AddOrMergeAsync(person);
+
+            var created = await entityTable.GetByIdAsync(person.TenantId, person.PersonId);
+
+            person.Enabled = !person.Enabled;
+
+            await entityTable.MergeAsync(person);
+
+            var merged = await entityTable.GetByIdAsync(person.TenantId, person.PersonId);
+
+            created.Should().BeEquivalentTo(merged, options => options.Excluding(p => p.Enabled));
+        }
+
+        [TestMethod]
+        public async Task Should_Partially_Merge_Entity()
+        {
+            var persons = Fakers.CreateFakePerson().Generate(1);
+            var person = persons.First();
+
+            var entityTable = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString)
+                .Configure(options => _defaultOptions(options), c =>
+                {
+                    c.
+                     SetPartitionKey(p => p.TenantId)
+                    .SetRowKeyProp(p => p.PersonId)
+                    .AddTag(p => p.LastName)
+                    .AddTag(p => p.Created);
+                });
+            await entityTable.AddOrMergeAsync(person);
+
+            var created = await entityTable.GetByIdAsync(person.TenantId, person.PersonId);
+
+            var personToMerge = new PersonEntity()
+            {
+                PersonId = person.PersonId,
+                TenantId = person.TenantId,
+                FirstName = "updatedFirstName",
+            };
+
+            await entityTable.MergeAsync(personToMerge);
+
+            var merged = await entityTable.GetByIdAsync(person.TenantId, person.PersonId);
+
+            merged.Should().NotBeNull();
+            merged.FirstName.Should().Be("updatedFirstName");
+        }
+
+        [TestMethod]
+        public async Task Should_Partially_Delete_Entity_After_Partial_Merge()
+        {
+            var persons = Fakers.CreateFakePerson().Generate(1);
+            var person = persons.First();
+
+            var entityTable = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString)
+                .Configure(options => _defaultOptions(options), c =>
+                {
+                    c.
+                     SetPartitionKey(p => p.TenantId)
+                    .SetRowKeyProp(p => p.PersonId)
+                    .AddTag(p => p.LastName)
+                    .AddTag(p => p.Created);
+                });
+            await entityTable.AddOrMergeAsync(person);
+
+            var personToMerge = new PersonEntity()
+            {
+                PersonId = person.PersonId,
+                TenantId = person.TenantId
+            };
+
+            await entityTable.MergeAsync(personToMerge);
+
+            await entityTable.DeleteByIdAsync(person.TenantId, person.PersonId);
+
+            var deleted = await entityTable.GetByIdAsync(person.TenantId, person.PersonId);
+
+            deleted.Should().BeNull();
+        }
+
+        [TestMethod]
+        public async Task Should_DeleteById_Entity()
+        {
+            var persons = Fakers.CreateFakePerson().Generate(1);
+            var person = persons.First();
+
+            var entityTable = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString)
+                .Configure(options => _defaultOptions(options), c =>
+                {
+                    c.
+                     SetPartitionKey(p => p.TenantId)
+                    .SetRowKeyProp(p => p.PersonId)
+                    .AddTag(p => p.LastName)
+                    .AddTag(p => p.Created);
+                });
+            await entityTable.AddOrMergeAsync(person);
+
+            var created = await entityTable.GetByIdAsync(person.TenantId, person.PersonId);
+
+            person.Enabled = !person.Enabled;
+
+            await entityTable.MergeAsync(person);
+
+            await entityTable.DeleteByIdAsync(person.TenantId, person.PersonId);
+            var deleted = await entityTable.GetByIdAsync(person.TenantId, person.PersonId);
+
+            deleted.Should().BeNull();
+        }
+
+        [TestMethod]
+        public async Task Should_DeleteBy_Entity()
+        {
+            var persons = Fakers.CreateFakePerson().Generate(1);
+            var person = persons.First();
+
+            var entityTable = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString)
+                .Configure(options => _defaultOptions(options), c =>
+                {
+                    c.
+                     SetPartitionKey(p => p.TenantId)
+                    .SetRowKeyProp(p => p.PersonId)
+                    .AddTag(p => p.LastName)
+                    .AddTag(p => p.Created);
+                });
+            await entityTable.AddOrMergeAsync(person);
+
+            var created = await entityTable.GetByIdAsync(person.TenantId, person.PersonId);
+
+            person.Enabled = !person.Enabled;
+
+            await entityTable.MergeAsync(person);
+
+            await entityTable.DeleteAsync(person);
+
+            var deleted = await entityTable.GetByIdAsync(person.TenantId, person.PersonId);
+
+            deleted.Should().BeNull();
+        }
+
+        [TestMethod]
+        public async Task Should_Update_Tag_When_Value_Updated()
+        {
+            var persons = Fakers.CreateFakePerson().Generate(1);
+            var person = persons.First();
+
+            var entityTable = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString)
+                .Configure(options =>
+                {
+                    _defaultOptions(options);
+                    options.HandleTagMutation = true;
+                }
+                , c =>
             {
                 c.
                  SetPartitionKey(p => p.TenantId)
@@ -74,6 +228,7 @@ namespace Azure.EntityServices.Table.Tests
 
             var created = await entityTable.GetByIdAsync(person.TenantId, person.PersonId);
 
+            created.Should().NotBeNull();
             created.LastName.Should().BeEquivalentTo(person.LastName);
 
             var tagResult = await entityTable.GetAsync(f => f.WhereTag(p => p.LastName)
@@ -97,7 +252,7 @@ namespace Azure.EntityServices.Table.Tests
             person.Altitude = null;
             var entityTable = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString)
                 .Configure(
-                options => _commonOptions(options),
+                options => _defaultOptions(options),
                 config =>
              {
                  config.
@@ -119,7 +274,7 @@ namespace Azure.EntityServices.Table.Tests
             var persons = Fakers.CreateFakePerson().Generate(1);
             var person = persons.First();
 
-            var entityTable = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _commonOptions(options), c =>
+            var entityTable = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _defaultOptions(options), c =>
             {
                 c.
                  SetPartitionKey(p => p.TenantId)
@@ -141,7 +296,7 @@ namespace Azure.EntityServices.Table.Tests
         {
             var persons = Fakers.CreateFakePerson().Generate(10);
 
-            var entityTable = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _commonOptions(options), c =>
+            var entityTable = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _defaultOptions(options), c =>
             {
                 c.
                 SetPartitionKey(p => p.TenantId)
@@ -172,7 +327,7 @@ namespace Azure.EntityServices.Table.Tests
         {
             var persons = Fakers.CreateFakePerson().Generate(10);
 
-            var entityTable = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _commonOptions(options), c =>
+            var entityTable = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _defaultOptions(options), c =>
             {
                 c.
                 SetPartitionKey(p => p.TenantId)
@@ -202,7 +357,7 @@ namespace Azure.EntityServices.Table.Tests
         public async Task Should_Set_Primary_Key_On_InsertOrUpdate()
         {
             var person = Fakers.CreateFakePerson().Generate();
-            var tableEntity = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _commonOptions(options), c =>
+            var tableEntity = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _defaultOptions(options), c =>
             {
                 c.SetPartitionKey(p => p.TenantId);
                 c.SetRowKeyProp(p => p.PersonId);
@@ -217,7 +372,7 @@ namespace Azure.EntityServices.Table.Tests
         public async Task Should_Get_Indexed_Tag_After_InsertOrUpdate()
         {
             var person = Fakers.CreateFakePerson().Generate();
-            var tableEntity = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _commonOptions(options), c =>
+            var tableEntity = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _defaultOptions(options), c =>
             {
                 c.SetPartitionKey(p => p.TenantId);
                 c.SetRowKeyProp(p => p.PersonId);
@@ -241,7 +396,7 @@ namespace Azure.EntityServices.Table.Tests
             static string First3Char(string s) => s.ToLower()[..3];
 
             var person = Fakers.CreateFakePerson().Generate();
-            var tableEntity = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _commonOptions(options), c =>
+            var tableEntity = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _defaultOptions(options), c =>
             {
                 c.SetPartitionKey(p => p.TenantId);
                 c.SetRowKeyProp(p => p.PersonId);
@@ -259,7 +414,7 @@ namespace Azure.EntityServices.Table.Tests
             static string First3Char(string s) => s.ToLower()[..3];
             var person = Fakers.CreateFakePerson().Generate();
             var tableEntity = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString)
-            .Configure(options => _commonOptions(options), c =>
+            .Configure(options => _defaultOptions(options), c =>
             {
                 c.SetPartitionKey(p => p.TenantId);
                 c.SetRowKeyProp(p => p.PersonId);
@@ -287,7 +442,7 @@ namespace Azure.EntityServices.Table.Tests
             //arrange entity
             var CreatedDate = DateTimeOffset.UtcNow;
             //
-            var tableEntity = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _commonOptions(options), c =>
+            var tableEntity = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _defaultOptions(options), c =>
             {
                 c.SetPartitionKey(p => p.TenantId);
                 c.SetRowKeyProp(p => p.PersonId);
@@ -308,7 +463,7 @@ namespace Azure.EntityServices.Table.Tests
             var CreatedDate = DateTimeOffset.UtcNow;
             person.Created = null;
 
-            var tableEntity = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _commonOptions(options), c =>
+            var tableEntity = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _defaultOptions(options), c =>
             {
                 c.SetPartitionKey(p => p.TenantId);
                 c.SetRowKeyProp(p => p.PersonId);
@@ -328,7 +483,7 @@ namespace Azure.EntityServices.Table.Tests
             //arrange entity
 
             var tableEntity = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString)
-            .Configure(options => _commonOptions(options), c =>
+            .Configure(options => _defaultOptions(options), c =>
             {
                 c.SetPartitionKey(p => p.TenantId);
                 c.SetRowKeyProp(p => p.PersonId);
@@ -365,7 +520,7 @@ namespace Azure.EntityServices.Table.Tests
 
             var person = Fakers.CreateFakePerson().Generate();
 
-            var tableEntity = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _commonOptions(options), c =>
+            var tableEntity = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _defaultOptions(options), c =>
             {
                 c.SetPartitionKey(p => p.TenantId);
                 c.SetRowKeyProp(p => p.PersonId);
@@ -406,7 +561,7 @@ namespace Azure.EntityServices.Table.Tests
             var persons = Fakers.CreateFakePerson().Generate(10);
             var observer = new DummyObserver();
 
-            var tableEntity = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _commonOptions(options), c =>
+            var tableEntity = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _defaultOptions(options), c =>
             {
                 c.SetPartitionKey(p => p.TenantId)
                 .SetRowKeyProp(p => p.PersonId)
@@ -432,7 +587,7 @@ namespace Azure.EntityServices.Table.Tests
             var partitionName = Guid.NewGuid().ToString();
             persons.ForEach(p => p.TenantId = partitionName);
 
-            IEntityTableClient<PersonEntity> tableEntity = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _commonOptions(options), config =>
+            IEntityTableClient<PersonEntity> tableEntity = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _defaultOptions(options), config =>
             {
                 config
                 .SetPartitionKey(p => p.TenantId)
@@ -462,7 +617,7 @@ namespace Azure.EntityServices.Table.Tests
             var partitionName = Guid.NewGuid().ToString();
             persons.ForEach(p => p.TenantId = partitionName);
 
-            IEntityTableClient<PersonEntity> tableEntity = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _commonOptions(options), config =>
+            IEntityTableClient<PersonEntity> tableEntity = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _defaultOptions(options), config =>
             {
                 config
                 .SetPartitionKey(p => p.TenantId)
@@ -499,7 +654,7 @@ namespace Azure.EntityServices.Table.Tests
             var partitionName = Guid.NewGuid().ToString();
             persons.ForEach(p => p.TenantId = partitionName);
 
-            IEntityTableClient<PersonEntity> tableEntity = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _commonOptions(options), config =>
+            IEntityTableClient<PersonEntity> tableEntity = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _defaultOptions(options), config =>
             {
                 config
                 .SetPartitionKey(p => p.TenantId)
@@ -517,353 +672,6 @@ namespace Azure.EntityServices.Table.Tests
         }
 
         [TestMethod]
-        public async Task Should_Filter_Entities()
-        {
-            var persons = Fakers.CreateFakePerson().Generate(130);
-
-            //force entities to have same partition (tenantId)
-            var partitionName = Guid.NewGuid().ToString();
-            persons.ForEach(p => p.TenantId = partitionName);
-
-            IEntityTableClient<PersonEntity> tableEntity = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _commonOptions(options), config =>
-            {
-                config
-                .SetPartitionKey(p => p.TenantId)
-                .SetRowKeyProp(p => p.PersonId)
-                .AddTag(p => p.LastName)
-                .AddTag(p => p.Created);
-            });
-
-            var latestPerson = persons.Last();
-            latestPerson.Altitude = -100;
-            await tableEntity.AddManyAsync(persons);
-            //get all entities both primary and projected
-            await foreach (var pagedResult in tableEntity.GetAsync(
-            filter => filter
-            .WherePartitionKey()
-            .Equal(persons.First().TenantId)
-            .And(p => p.Altitude)
-            .Equal(latestPerson.Altitude)))
-            {
-                pagedResult.Should().HaveCount(1);
-            }
-        }
-
-        //for now query with null values are not supported in Azure Table Storage
-        //event if, it works in Storage Emulator
-        //[TestMethod]
-        public async Task Should_Filter_Entities_With_Nullable_Properties()
-        {
-            var persons = Fakers.CreateFakePerson().Generate(10);
-
-            //force entities to have same partition (tenantId)
-            var partitionName = Guid.NewGuid().ToString();
-            persons.ForEach(p => p.TenantId = partitionName);
-
-            IEntityTableClient<PersonEntity> tableEntity = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _commonOptions(options), config =>
-            {
-                config
-                .SetPartitionKey(p => p.TenantId)
-                .SetRowKeyProp(p => p.PersonId)
-                .AddTag(p => p.LastName)
-                .AddTag(p => p.Created);
-            });
-
-            var latestPerson = persons.Last();
-            latestPerson.BankAmount = null;
-            latestPerson.Altitude = null;
-            latestPerson.Situation = null;
-            latestPerson.Enabled = true;
-            await tableEntity.AddManyAsync(persons);
-            //get all entities both primary and projected
-            await foreach (var pagedResult in tableEntity.GetAsync(
-            filter => filter
-            .WherePartitionKey()
-                .Equal(persons.First().TenantId)
-                .And(p => p.BankAmount)
-                .Equal(null)
-                .And(p => p.Altitude)
-                .Equal(null)
-                .And(p => p.Situation)
-                .Equal(null)
-                .And(p => p.Enabled)
-                .NotEqual(null))
-                )
-            {
-                pagedResult.Should().HaveCount(1);
-                pagedResult.First().Should().BeEquivalentTo(latestPerson);
-            }
-        }
-
-        [TestMethod]
-        public async Task Should_Query_By_Tag_Filter_With_Equal_Extension()
-        {
-            var persons = Fakers.CreateFakePerson().Generate(10);
-
-            var entityTable = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _commonOptions(options), c =>
-            {
-                c.
-                SetPartitionKey(p => p.TenantId)
-                .SetRowKeyProp(p => p.PersonId)
-                .AddTag(p => p.LastName)
-                .AddTag(p => p.Created);
-            });
-
-            await entityTable.AddManyAsync(persons);
-
-            var person = persons.Last();
-
-            await foreach (var resultPage in entityTable.GetAsync(
-            filter => filter
-            .WhereTag(p => p.Created)
-            .Equal(person.Created)))
-            {
-                resultPage.First().Should().BeEquivalentTo(person);
-            }
-        }
-
-        [TestMethod]
-        public async Task Should_Query_By_Tag_Filter_With_GreaterThanOrEqual()
-        {
-            var persons = Fakers.CreateFakePerson().Generate(10);
-
-            var entityTable = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _commonOptions(options), c =>
-            {
-                c.
-                SetPartitionKey(p => p.TenantId)
-                .SetRowKeyProp(p => p.PersonId)
-                .AddTag(p => p.LastName)
-                .AddTag(p => p.Created);
-            });
-
-            var oldestDate = persons.Min(p => p.Created.Value);
-            var olderPerson = Fakers.CreateFakePerson().Generate(1).First();
-            olderPerson.Created = oldestDate - TimeSpan.FromSeconds(1);
-            persons.Add(olderPerson);
-            await entityTable.AddManyAsync(persons);
-
-            await foreach (var resultPage in entityTable.GetAsync(filter => filter.WhereTag("Created").GreaterThanOrEqual(oldestDate)))
-            {
-                resultPage.Should().BeEquivalentTo(persons.Where(p => p != olderPerson));
-            }
-        }
-
-        [TestMethod]
-        public async Task Should_Query_By_Tag_Filter_With_GreaterThan()
-        {
-            var persons = Fakers.CreateFakePerson().Generate(10);
-
-            var entityTable = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _commonOptions(options), c =>
-            {
-                c.
-                SetPartitionKey(p => p.TenantId)
-                .SetRowKeyProp(p => p.PersonId)
-                .AddTag(p => p.LastName)
-                .AddTag(p => p.Created);
-            });
-
-            var oldestDate = persons.Min(p => p.Created.Value);
-            var olderPerson = Fakers.CreateFakePerson().Generate(1).First();
-            olderPerson.Created = oldestDate - TimeSpan.FromSeconds(1);
-            persons.Add(olderPerson);
-            await entityTable.AddManyAsync(persons);
-
-            await foreach (var resultPage in entityTable.GetAsync(
-                filter => filter
-                .WhereTag("Created").GreaterThan(oldestDate)))
-            {
-                resultPage.Select(p => p.Created).All(p => p.Value > oldestDate).Should().BeTrue();
-                resultPage.Should().NotContain(olderPerson);
-            }
-        }
-
-        [TestMethod]
-        public async Task Should_Query_By_Tag_Filter_With_LessThan()
-        {
-            var persons = Fakers.CreateFakePerson().Generate(10);
-
-            var entityTable = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _commonOptions(options), c =>
-            {
-                c.
-                SetPartitionKey(p => p.TenantId)
-                .SetRowKeyProp(p => p.PersonId)
-                .AddTag(p => p.LastName)
-                .AddTag(p => p.Created);
-            });
-
-            var latestDate = persons.Max(p => p.Created);
-            var lastestPerson = Fakers.CreateFakePerson().Generate(1).First();
-            lastestPerson.Created = latestDate + TimeSpan.FromSeconds(1);
-            persons.Add(lastestPerson);
-            await entityTable.AddManyAsync(persons);
-
-            await foreach (var resultPage in entityTable.GetAsync(
-                filter => filter
-                .WhereTag(p => p.Created)
-                .LessThan(latestDate)))
-            {
-                resultPage.Select(p => p.Created).All(p => p.Value < latestDate).Should().BeTrue();
-                resultPage.Should().NotContain(lastestPerson);
-            }
-        }
-
-        [TestMethod]
-        public async Task Should_Query_By_Tag_Filter_With_LessThanOrEqual()
-        {
-            var persons = Fakers.CreateFakePerson().Generate(10);
-
-            var entityTable = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _commonOptions(options), c =>
-            {
-                c.
-                SetPartitionKey(p => p.TenantId)
-                .SetRowKeyProp(p => p.PersonId)
-                .AddTag(p => p.LastName)
-                .AddTag(p => p.Created);
-            });
-
-            var latestDate = persons.Max(p => p.Created);
-            var lastestPerson = Fakers.CreateFakePerson().Generate(1).First();
-            lastestPerson.Created = latestDate + TimeSpan.FromSeconds(1);
-            persons.Add(lastestPerson);
-            await entityTable.AddManyAsync(persons);
-            await foreach (var resultPage in entityTable.GetAsync(
-
-                filter => filter
-                .WhereTag("Created")
-                .LessThanOrEqual(latestDate)))
-            {
-                resultPage.Should().BeEquivalentTo(persons.Where(p => p != lastestPerson));
-            }
-        }
-
-        [TestMethod]
-        public async Task Should_Query_By_Tag_Filter_With_Between_Extension()
-        {
-            var persons = Fakers.CreateFakePerson().Generate(10);
-
-            var entityTable = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _commonOptions(options), c =>
-            {
-                c.
-                SetPartitionKey(p => p.TenantId)
-                .SetRowKeyProp(p => p.PersonId)
-                .AddTag(p => p.LastName)
-                .AddTag(p => p.Created);
-            });
-
-            var latestDate = persons.Max(p => p.Created);
-            var latestPerson = Fakers.CreateFakePerson().Generate(1).First();
-            latestPerson.Created = latestDate + TimeSpan.FromSeconds(1);
-            persons.Add(latestPerson);
-
-            var oldestDate = persons.Min(p => p.Created.Value);
-            var olderPerson = Fakers.CreateFakePerson().Generate(1).First();
-            olderPerson.Created = oldestDate - TimeSpan.FromSeconds(1);
-            persons.Add(olderPerson);
-
-            await entityTable.AddManyAsync(persons);
-            await foreach (var resultPage in entityTable.GetAsync(
-                filter =>
-                filter
-                .WhereTag(p => p.Created)
-                .Between(oldestDate, latestDate)))
-            {
-                resultPage.Should().BeEquivalentTo(persons
-                    .Where(p => p.Created > olderPerson.Created
-                    &&
-                    p.Created < latestPerson.Created));
-            }
-        }
-
-        [TestMethod]
-        public async Task Should_Combine_Query_With_Mixed_Tag_And_Prop_Filters()
-        {
-            var persons = Fakers.CreateFakePerson().Generate(10);
-
-            var entityTable = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _commonOptions(options), c =>
-            {
-                c.
-                SetPartitionKey(p => p.TenantId)
-                .SetRowKeyProp(p => p.PersonId)
-                .AddTag(p => p.LastName)
-                .AddTag(p => p.Created);
-            });
-
-            persons.Last().Created = DateTimeOffset.UtcNow;
-            persons.Last().Altitude = -100;
-
-            await entityTable.AddManyAsync(persons);
-            //Query by Created Tag
-            await foreach (var resultPage in entityTable.GetAsync(
-                filter =>
-                    filter
-                    .WhereTag(p => p.Created)
-                    .GreaterThanOrEqual(persons.Last().Created)
-                    .And(p => p.Altitude)
-                    .LessThanOrEqual(-100)
-
-                ))
-            {
-                resultPage.Count().Should().BePositive();
-                resultPage.Select(p => p.Altitude).All(p => p <= 100).Should().BeTrue();
-            }
-        }
-
-        [TestMethod]
-        public async Task Should_Query_By_Tag_Filter_With_Nullable_Values()
-        {
-            var persons = Fakers.CreateFakePerson().Generate(10);
-
-            var entityTable = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _commonOptions(options), c =>
-            {
-                c.
-                SetPartitionKey(p => p.TenantId)
-                .SetRowKeyProp(p => p.PersonId)
-                .AddTag(p => p.LastName)
-                .AddTag(p => p.Created);
-            });
-
-            var person = persons.Last();
-            person.Created = null;
-
-            await entityTable.AddManyAsync(persons);
-
-            await foreach (var resultPage in entityTable.GetAsync(
-            filter => filter
-            .WhereTag(p => p.Created)
-            .Equal(null)))
-            {
-                resultPage.First().Should().BeEquivalentTo(person);
-            }
-        }
-
-        [TestMethod]
-        public async Task Should_Query_By_Tag_Filter_With_Empty_Values()
-        {
-            var persons = Fakers.CreateFakePerson().Generate(1);
-
-            var entityTable = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _commonOptions(options), c =>
-            {
-                c.
-                SetPartitionKey(p => p.TenantId)
-                .SetRowKeyProp(p => p.PersonId)
-                .AddTag(p => p.LastName)
-                .AddTag(p => p.Created);
-            });
-
-            var person = persons.Last();
-            person.LastName = string.Empty;
-
-            await entityTable.AddManyAsync(persons);
-            await foreach (var resultPage in entityTable.GetAsync(
-            filter => filter
-            .WhereTag(p => p.LastName)
-            .Equal("")))
-            {
-                resultPage.First().Should().BeEquivalentTo(person);
-            }
-        }
-
-        [TestMethod]
         public async Task Should_Store_Default_DateTime_Values()
         {
             var person = Fakers.CreateFakePerson().Generate(1).FirstOrDefault();
@@ -873,7 +681,7 @@ namespace Azure.EntityServices.Table.Tests
             person.LocalCreated = default;
             person.LocalUpdated = default;
 
-            var entityTable = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _commonOptions(options), c =>
+            var entityTable = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _defaultOptions(options), c =>
             {
                 c
                 .SetPartitionKey(p => p.TenantId)
@@ -895,7 +703,7 @@ namespace Azure.EntityServices.Table.Tests
             person.Created = null;
             person.LocalCreated = null;
 
-            var entityTable = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _commonOptions(options), c =>
+            var entityTable = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _defaultOptions(options), c =>
             {
                 c
                 .SetPartitionKey(p => p.TenantId)
@@ -918,7 +726,7 @@ namespace Azure.EntityServices.Table.Tests
             person.Created = null;
             person.LocalCreated = null;
 
-            var entityTable = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _commonOptions(options), c =>
+            var entityTable = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _defaultOptions(options), c =>
             {
                 c
                 .SetPartitionKey(p => p.TenantId)
@@ -944,7 +752,7 @@ namespace Azure.EntityServices.Table.Tests
             person.Created = null;
             person.LocalCreated = null;
 
-            var entityTable = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _commonOptions(options), c =>
+            var entityTable = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString).Configure(options => _defaultOptions(options), c =>
             {
                 c
                 .SetPartitionKey(p => p.TenantId)
@@ -971,7 +779,7 @@ namespace Azure.EntityServices.Table.Tests
         {
             var person = Fakers.CreateFakePerson().Generate(1).FirstOrDefault();
             var options = new EntityTableClientOptions() { };
-            _commonOptions.Invoke(options);
+            _defaultOptions.Invoke(options);
             options.ConfigureSerializerWithStringEnumConverter();
 
             var genericClient = EntityTableClient.Create<TableEntity>(TestEnvironment.ConnectionString)
@@ -1009,11 +817,11 @@ namespace Azure.EntityServices.Table.Tests
         }
 
         [TestMethod]
-        public async Task Should_DeSerialize_Existing_Enum_As_Integer()
+        public async Task Should_Deserialize_Existing_Enum_As_Integer()
         {
             var person = Fakers.CreateFakePerson().Generate(1).FirstOrDefault();
             var options = new EntityTableClientOptions() { };
-            _commonOptions.Invoke(options);
+            _defaultOptions.Invoke(options);
 
             var genericClient = EntityTableClient.Create<TableEntity>(TestEnvironment.ConnectionString)
                 .Configure(options, c =>
@@ -1048,169 +856,6 @@ namespace Azure.EntityServices.Table.Tests
             var updatedEntity = await personClient.GetByIdAsync(person.TenantId, person.LastName);
             updatedEntity.Address.Should()
                 .BeEquivalentTo(person.Address);
-        }
-
-        [TestMethod]
-        public async Task Should_Get_Paged_Entities()
-        {
-            var persons = Fakers.CreateFakePerson().Generate(155);
-            var options = new EntityTableClientOptions() { };
-            _commonOptions.Invoke(options);
-
-            var personClient = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString)
-                .Configure(options,
-            c =>
-            {
-                c
-                .SetPartitionKey(p => p.TenantId)
-                .SetRowKeyProp(p => p.LastName)
-                .AddTag(p => p.LastName)
-                .AddTag(p => p.Created);
-            });
-
-            await personClient.AddManyAsync(persons);
-
-            EntityPage<PersonEntity> page = default;
-            int entityCount = 0;
-
-            do
-            {
-                page = await personClient.GetPagedAsync(nextPageToken: page.ContinuationToken);
-
-                entityCount += page.Entities.Count();
-            }
-            while (!page.IsLastPage);
-
-            entityCount.Should().Be(155);
-        }
-
-        //[DataRow[ available_entities, max_per_page ]
-        [DataRow(10, 20)] //skip more than available entities
-        [DataRow(158, 100)]
-        [DataRow(1024, 999)]
-        [DataRow(2012, 1000)]
-        [TestMethod]
-        public async Task Should_Get_Paged_Entities_With_Custom_Max_Per_Page(params int[] inputs)
-        {
-            int totalCount = inputs[0];
-            int maxPerPage = inputs[1];
-
-            var persons = Fakers.CreateFakePerson().Generate(totalCount);
-            var options = new EntityTableClientOptions() { };
-            _commonOptions.Invoke(options);
-
-            var personClient = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString)
-                .Configure(options,
-            c =>
-            {
-                c
-                .SetPartitionKey(p => p.TenantId)
-                .SetRowKeyProp(p => p.LastName)
-                .AddTag(p => p.LastName)
-                .AddTag(p => p.Created);
-            });
-
-            await personClient.AddManyAsync(persons);
-
-            EntityPage<PersonEntity> page = default;
-            int entityCount = 0;
-
-            do
-            {
-                page = await personClient.GetPagedAsync(maxPerPage: maxPerPage, iteratedCount: page.IteratedCount, nextPageToken: page.ContinuationToken);
-
-                if (page.IsLastPage)
-                {
-                    page.Entities.Count().Should().BeLessThanOrEqualTo(maxPerPage);
-                }
-                else
-                {
-                    page.Entities.Count().Should().Be(maxPerPage);
-                }
-                entityCount += page.Entities.Count();
-            }
-            while (!page.IsLastPage);
-
-            entityCount.Should().Be(totalCount);
-            page.IteratedCount.Should().Be(entityCount);
-        }
-
-        //[DataRow[ available_entities, to_skip]
-        [DataRow(10, 10)] //skip more than available entities
-        [DataRow(158, 10)]
-        [DataRow(122, 5)]
-        [DataRow(1024, 500)]
-        [DataRow(2012, 1200)]
-        [TestMethod]
-        public async Task Should_Skip_Entities_With_AsyncEnumerableExtensions(params int[] inputs)
-        {
-            int totalCount = inputs[0];
-            int skipCount = inputs[1];
-
-            var persons = Fakers.CreateFakePerson().Generate(totalCount);
-            var options = new EntityTableClientOptions() { };
-            _commonOptions.Invoke(options);
-
-            var personClient = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString)
-                .Configure(options,
-            c =>
-            {
-                c
-                .SetPartitionKey(p => p.TenantId)
-                .SetRowKeyProp(p => p.LastName)
-                .AddTag(p => p.LastName)
-                .AddTag(p => p.Created);
-            });
-
-            await personClient.AddManyAsync(persons);
-
-            var all = await personClient.GetAsync().ToListAsync();
-
-            var skipResult = await personClient
-                .GetAsync()
-                .SkipAsync(skipCount)
-                .ToListAsync();
-
-            skipResult.Count.Should().Be(Math.Max(0, totalCount - skipCount));
-        }
-
-        //[DataRow[ available_entities, to_take ]
-        [DataRow(10, 20)] //skip more than available entities
-        [DataRow(158, 20)]
-        [DataRow(122, 100)]
-        [DataRow(1024, 999)]
-        [DataRow(2012, 1000)]
-        [TestMethod]
-        public async Task Should_Take_Entities_With_AsyncEnumerableExtensions(params int[] inputs)
-        {
-            int totalCount = inputs[0];
-            int takeCount = inputs[1];
-
-            var persons = Fakers.CreateFakePerson().Generate(totalCount);
-            var options = new EntityTableClientOptions() { };
-            _commonOptions.Invoke(options);
-
-            var personClient = EntityTableClient.Create<PersonEntity>(TestEnvironment.ConnectionString)
-                .Configure(options,
-            c =>
-            {
-                c
-                .SetPartitionKey(p => p.TenantId)
-                .SetRowKeyProp(p => p.LastName)
-                .AddTag(p => p.LastName)
-                .AddTag(p => p.Created);
-            });
-
-            await personClient.AddManyAsync(persons);
-
-            var all = await personClient.GetAsync().ToListAsync();
-
-            var takeResult = await personClient
-                .GetAsync()
-                .TakeAsync(takeCount)
-                .ToListAsync();
-
-            takeResult.Count.Should().Be(Math.Min(totalCount, takeCount));
         }
     }
 }
